+++ tests/oidc/test_consent.py
@@ -4,13 +4,13 @@ def test_no_logged_no_access(testclient):
 
 def test_consent_list(testclient, client, consent, logged_user, token):
     res = testclient.get("/consent", status=200)
-    assert client.name in res.text
+    assert client.client_name in res.text
     assert not token.revoked
 
     res = testclient.get(f"/consent/delete/{consent.cn[0]}", status=302)
 
     res = res.follow(status=200)
-    assert client.name not in res.text
+    assert client.client_name not in res.text
 
     token.reload()
     assert token.revoked