+++ tests/oidc/test_password_flow.py
@@ -48,7 +48,7 @@ def test_password_flow_post(testclient, user, client):
             password="correct horse battery staple",
             scope="openid profile groups",
             client_id=client.client_id,
-            client_secret=client.secret,
+            client_secret=client.client_secret,
         ),
         status=200,
     )